refactor: extract CompactionGuards service from LcmContextEngine
Phase 2 of the engine.ts decomposition (#867). The circuit breaker for
compaction auth failures and the per-scope summary spend guard move from
private engine methods + two private Maps into src/compaction-guards.ts
(CompactionGuards), constructed with config + deps. Method bodies are
unchanged; resolvePositiveConfigInteger becomes the shared
resolvePositiveInteger helper in value-utils. Tests that reached into the
old private methods now go through the compactionGuards field.

Full suite: 1279 tests passing.
